aarch64: whitelist necessary syscalls
Moved common functionality in mod.rs and left the only
architectural specific function (i.e default_filter) inside
x86_64.rs and aarch64.rs, respectively.
